refactor(service): split core into separate packages
